Refactor to add `create_area_def_from_dict` and `create_area_def_from_yaml` utility functions
